Added Loan Modules
Create loan / Update loans

+++ database/migrations/2020_03_04_023358_create_household_incomes_table.php
@@ -22,7 +22,7 @@ class CreateHouseholdIncomesTable extends Migration
             $table->boolean('is_employed');
             $table->string('employed_position')->nullable();
             $table->string('employed_company_name')->nullable();
-            $table->integer('employed_monthly_gross_income')->nullable();
+            $table->double('employed_monthly_gross_income')->nullable();
 
             //spouse
             $table->boolean('spouse_is_self_employed');
@@ -31,15 +31,15 @@ class CreateHouseholdIncomesTable extends Migration
             $table->boolean('spouse_is_employed');
             $table->string('spouse_employed_position')->nullable();
             $table->string('spouse_employed_company_name')->nullable();
-            $table->integer('spouse_employed_monthly_gross_income')->nullable();
+            $table->double('spouse_employed_monthly_gross_income')->nullable();
 
             $table->boolean('has_remittance')->default(false);
-            $table->integer('remittance_amount')->nullable();
+            $table->double('remittance_amount')->nullable();
 
             $table->boolean('has_pension')->default(false);
-            $table->integer('pension_amount')->nullable();
+            $table->double('pension_amount')->nullable();
 
-            $table->integer('total_household_income');
+            $table->double('total_household_income');
 
             $table->timestamps();
         });
