+++ database/migrations/2020_03_23_030847_create_deposits_table.php
@@ -22,9 +22,9 @@ class CreateDepositsTable extends Migration
             $table->date('valid_until')->nullable();
             $table->integer('account_per_client');
             $table->boolean('auto_create_on_new_client')->default(false);
-            $table->integer('interest_rate');
-            $table->integer('deposit_portfolio');
-            $table->integer('deposit_interest_expense');
+            $table->double('interest_rate');
+            $table->unsignedInteger('deposit_portfolio');
+            $table->unsignedInteger('deposit_interest_expense');
             $table->boolean('is_active')->default(true);
